requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.6.15
chardet==5.2.0
charset-normalizer==3.4.2
dj-database-url==3.0.1
dj-rest-auth==7.0.1
Django==5.2.3
django-allauth==65.9.0
django-axes==8.0.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==24.1
django-jazzmin==3.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
easyocr==1.7.2
et_xmlfile==2.0.0
filelock==3.19.1
fsspec==2025.9.0
gunicorn==21.2.0
idna==3.10
imageio==2.37.0
inflection==0.5.1
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
ninja==1.13.0
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pdf2image==1.17.0
pillow==11.2.1
psycopg2-binary==2.9.10
pyclipper==1.3.0.post6
PyJWT==2.9.0
pytesseract==0.3.13
python-bidi==0.6.6
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
reportlab==4.4.1
requests==2.32.4
rpds-py==0.25.1
scikit-image==0.25.2
scipy==1.16.1
shapely==2.1.1
sqlparse==0.5.3
sympy==1.14.0
tifffile==2025.8.28
torch==2.8.0
torchvision==0.23.0
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
drf-yasg==1.21.6

